// src/types/elements.ts
export type Rounded = 'full' | 'lg' | 'md' | 'none' | 'xl'

export type Color =
  | 'accent'
  | 'danger'
  | 'error'
  | 'info'
  | 'none'
  | 'primary'
  | 'secondary'
  | 'success'
  | 'warning'

export type Tone =
  | '50'
  | '100'
  | '200'
  | '300'
  | '400'
  | '500'
  | '600'
  | '700'
  | '800'
  | '900'
  | '950'

export type Font = 'bold' | 'normal' | 'semibold'

export type Spacing = 'lg' | 'md' | 'none' | 'sm' | 'xs'

export type Size =
  | '2xl'
  | '3xl'
  | '4xl'
  | '5xl'
  | 'lg'
  | 'md'
  | 'none'
  | 'sm'
  | 'xl'
  | 'xs'